UniversityRecordManager-Project/requirements.txt
django_bootstrap_modal_forms==1.3.1
Django==2.1.7
